UPDATE file_i_check.bats - add num test, fix name
Added a cheap "numerical" test (a test that inputs a number) that in
reality probably isn't worth much as all input would be evaluated as a
file, but overall there for sanity/regression testing purposes in case
the checks within the file_integrity_check() func are ever removed.

Also renamed another function slightly, to better reflect what it really
does.