UPDATE utils.sh - swap -z test for arg number test
Removed the original -z based test as it no longer had as much of an
impact as it turns out that I can do more succinct testing logic with
the $# expression, I can do exact bounds checking basically.

Also removed the explicit exit 0 from the code, as there's only one
other option and I felt it was cleaner to just have it perform its
natural/default exit 0 when it didn't encounter an issue.