UPDATE pstart - update used shell options
Removed the POSIX and nounset shell options from the script as they
induced errors when interacting with other shell scripts sourced, there
was an issue basically with the nounset option where a 3rd party script
was using an unset variable and my shell got into an infinite loop, with
the only option available to remove it. POSIX was removed because people
tend to not use it, hence it breaks things.

Add the noclobbet shell option to prevent errant io redirection from
nuking files. This was just a guardrail addition, basically.