UPDATE pstart - re implement create_makefile func
The create_gitignore func previously did not work, the use of the local
ignores variable list didn't work as I'd previously thought it would so
it was next to useless. I've replaced it with a 'declare array' that
allows for proper looping of the data structure and is actually part of
the bash syntax.

Overall, this commit gets the entire func to work, as previously it was
littered with semi pseudo code garbage.