UPDATE pstart - fix double func invocation bug
So I had a problem with the testing for a while that was pretty
infuriating and it turned out to be my little 'test hack' that I had
originally used to test my scrip that my create_gitignore func was
working, so I injected it into the bottom of my script so I could call
the main binary. I then forgot to remove it, and because what I was
testing was the file_integrity_check it kept 'failing' because the
previous run would create the gitignore, then the integrity check would
be called again by the second explicit invocation of the
create_gitignore that was being called by the tests, but the sourcing of
the main binary would implicitly invoke the func.