UPDATE Makefile - rewrite the install rule
The install rule was too flaky, it didn't handle situations where the directories already existed and basically it wouldn't update the dependencies in the utils.sh file. In reality, follow up commits will probably create a new rule that will be used just for updating development perms, as it
is unlikely that this rule will be used so much by 'normal' users. It's now contextual, allowing for changes in the utils.sh file to be picked up and pushed into the relevant lib folders