UPDATE pstart - add arg parsing and show_help func
This commit adds in very basic command line parsing so the program can
be called from the command line, currently only allowing the show_help
function to be called. Which is still very bare bones but in a working
state. Future commits will introduce more functionality to the callable
binary.